RT-32705: TextField/TextArea: on Windows, ArrayIndexOutOfBoundsException for complex Thai characters

+++ modules/graphics/src/main/java/com/sun/javafx/font/directwrite/JFXTextRenderer.java
@@ -68,7 +68,7 @@ class JFXTextRenderer extends IUnknown {
         return OS.JFXTextRendererGetGlyphOffsets(ptr, offsets, start);
     }
 
-    int GetClusterMap(int[] clusterMap, int start) {
+    int GetClusterMap(short[] clusterMap, int start) {
         return OS.JFXTextRendererGetClusterMap(ptr, clusterMap, start);
     }
 }
